+++ accounts/views_student.py
@@ -192,16 +192,29 @@ def edit_student_profile(request):
             student.pincode = encryptionHelper.encrypt(request.POST["pincode"])
 
             if request.FILES:
-                x = student.profile_pic.url.split("/account/media/")
-                if x[1] != "default.svg":
-                    file = settings.MEDIA_ROOT + "/" + x[1]
-                    os.remove(file)
-                student.profile_pic = request.FILES["profile_pic"]
+                if request.FILES["profile_pic"].size > 5 * 1024 * 1024:
+                    form.add_error(
+                        "profile_pic", "File size must be less than 5MB."
+                    )
+
+                    return render(
+                        request,
+                        "student/update_students_info.html",
+                        {
+                            "form": form,
+                        },
+                    )
+                else:
+                    x = student.profile_pic.url.split("/account/media/")
+                    if x[1] != "default.svg":
+                        file = settings.MEDIA_ROOT + '\\' + x[1]
+                        os.remove(file)
+                    student.profile_pic = request.FILES["profile_pic"]
             else:
                 if "profile_pic-clear" in request.POST.keys():
                     x = student.profile_pic.url.split("/account/media/")
                     if x[1] != "default.svg":
-                        file = settings.MEDIA_ROOT + "/" + x[1]
+                        file = settings.MEDIA_ROOT + '\\' + x[1]
                         os.remove(file)
                     student.profile_pic = "/default.svg"
