+++ accounts/views_supercoordinator.py
@@ -495,16 +495,29 @@ def edit_supercoordinator_profile(request):
             supercoordinator.gender = encryptionHelper.encrypt(request.POST["gender"])
 
             if request.FILES:
-                x = supercoordinator.profile_pic.url.split("/account/media/")
-                if x[1] != "default.svg":
-                    file = settings.MEDIA_ROOT + "/" + x[1]
-                    os.remove(file)
-                supercoordinator.profile_pic = request.FILES["profile_pic"]
+                if request.FILES["profile_pic"].size > 5 * 1024 * 1024:
+                    form.add_error(
+                        "profile_pic", "File size must be less than 5MB."
+                    )
+
+                    return render(
+                        request,
+                        "supercoordinator/update_supercoordinators_info.html",
+                        {
+                            "form": form,
+                        },
+                    )
+                else:
+                    x = supercoordinator.profile_pic.url.split("/account/media/")
+                    if x[1] != "default.svg":
+                        file = settings.MEDIA_ROOT + '\\' + x[1]
+                        os.remove(file)
+                    supercoordinator.profile_pic = request.FILES["profile_pic"]
             else:
                 if "profile_pic-clear" in request.POST.keys():
                     x = supercoordinator.profile_pic.url.split("/account/media/")
                     if x[1] != "default.svg":
-                        file = settings.MEDIA_ROOT + "/" + x[1]
+                        file = settings.MEDIA_ROOT + '\\' + x[1]
                         os.remove(file)
                     supercoordinator.profile_pic = "/default.svg"
