+++ accounts/views_teacher.py
@@ -1828,16 +1828,29 @@ def edit_teacher_profile(request):
             teacher.gender = encryptionHelper.encrypt(request.POST["gender"])
 
             if request.FILES:
-                x = teacher.profile_pic.url.split('/account/media/')
-                if x[1] != 'default.svg':
-                    file = settings.MEDIA_ROOT + '/' + x[1]
-                    os.remove(file)
-                teacher.profile_pic = request.FILES["profile_pic"]
+                if request.FILES["profile_pic"].size > 5 * 1024 * 1024:
+                    form.add_error(
+                        "profile_pic", "File size must be less than 5MB."
+                    )
+
+                    return render(
+                        request,
+                        "teacher/update_teachers_info.html",
+                        {
+                            "form": form,
+                        },
+                    )
+                else:
+                    x = teacher.profile_pic.url.split('/account/media/')
+                    if x[1] != 'default.svg':
+                        file = settings.MEDIA_ROOT + '\\' + x[1]
+                        os.remove(file)
+                    teacher.profile_pic = request.FILES["profile_pic"]
             else:
                 if "profile_pic-clear" in request.POST.keys():
                     x = teacher.profile_pic.url.split('/account/media/')
                     if x[1] != 'default.svg':
-                        file = settings.MEDIA_ROOT + '/' + x[1]
+                        file = settings.MEDIA_ROOT + '\\' + x[1]
                         os.remove(file)
                     teacher.profile_pic = "/default.svg"
